+++ src/constants.ts
@@ -8,12 +8,12 @@ export const StoreRepositoryStatus = {
 export type StoreRepositoryStatus =
 	(typeof StoreRepositoryStatus)[keyof typeof StoreRepositoryStatus]
 
-export const StoreRepositoryMethod = {
+export const StoreRepositoryAction = {
 	read: 'read',
 	create: 'create',
 	update: 'update',
 	remove: 'remove',
 } as const
 
-export type StoreRepositoryMethod =
-	(typeof StoreRepositoryMethod)[keyof typeof StoreRepositoryMethod]
+export type StoreRepositoryAction =
+	(typeof StoreRepositoryAction)[keyof typeof StoreRepositoryAction]